We did a redesign, but I regret using decimal instead of regular floats for prices.

+++ Spiral.Trading/F# version/Spiral.Trading/sql/schema/views/split_adjusted_prices.sql
@@ -36,5 +36,5 @@ SELECT
     high / adj_factor AS adj_high,
     low / adj_factor AS adj_low,
     close / adj_factor AS adj_close,
-    CAST(volume * adj_factor AS INTEGER) AS adj_volume
+    CAST(volume * adj_factor AS BIGINT) AS adj_volume
 FROM split_factors;
